Update example.mpv.conf
After fighting with stuff for a while, this is what ended up giving me the best results.

Some madmen use floyd-steinberg error dithering. It seems to work okay on 1080p, but will lag badly on 4k. It's VERY resource intensive and the gains from it are not going to be noticeable in mpv's use case.